requirements.txt
torch==1.8.0
torchvision==0.9.0
numpy==1.20
pyyaml==6.0